// frontend/src/api/generated/models/UpdateContractDetails.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateContractDetails
 */
export interface UpdateContractDetails {
    /**
     * 
     * @type {number}
     * @memberof UpdateContractDetails
     */
    contractId: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateContractDetails
     */
    roomId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateContractDetails
     */
    amount?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateContractDetails
     */
    deposit?: number | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateContractDetails
     */
    startDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateContractDetails
     */
    endDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateContractDetails
     */
    payedTillDayOfMonth?: string | null;
}

/**
 * Check if a given object implements the UpdateContractDetails interface.
 */
export function instanceOfUpdateContractDetails(value: object): value is UpdateContractDetails {
    if (!('contractId' in value) || value['contractId'] === undefined) return false;
    return true;
}

export function UpdateContractDetailsFromJSON(json: any): UpdateContractDetails {
    return UpdateContractDetailsFromJSONTyped(json, false);
}

export function UpdateContractDetailsFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateContractDetails {
    if (json == null) {
        return json;
    }
    return {
        
        'contractId': json['contractId'],
        'roomId': json['roomId'] == null ? undefined : json['roomId'],
        'amount': json['amount'] == null ? undefined : json['amount'],
        'deposit': json['deposit'] == null ? undefined : json['deposit'],
        'startDate': json['startDate'] == null ? undefined : json['startDate'],
        'endDate': json['endDate'] == null ? undefined : json['endDate'],
        'payedTillDayOfMonth': json['payedTillDayOfMonth'] == null ? undefined : json['payedTillDayOfMonth'],
    };
}

export function UpdateContractDetailsToJSON(value?: UpdateContractDetails | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'contractId': value['contractId'],
        'roomId': value['roomId'],
        'amount': value['amount'],
        'deposit': value['deposit'],
        'startDate': value['startDate'],
        'endDate': value['endDate'],
        'payedTillDayOfMonth': value['payedTillDayOfMonth'],
    };
}